src/requirements.txt
numpy==2.3.1
opencv_python==4.11.0.86
pandas==2.3.0
Pillow==11.2.1
timm==1.0.15
torch==2.7.1
torchvision==0.22.1

streamlit
xgboost
scikit-learn
imbalanced-learn
matplotlib
seaborn
joblib

